+++ src/main/java/de/codescape/jira/plugins/scrumpoker/action/ScrumPokerProjectConfigurationAction.java
@@ -97,7 +97,7 @@ public class ScrumPokerProjectConfigurationAction extends AbstractScrumPokerActi
      */
     public ProjectSettings getProjectSettings() {
         ScrumPokerProject scrumPokerProject = projectSettingsService.loadSettings(getProjectByKey(projectKey).getId());
-        return scrumPokerProject != null ? new ProjectSettings(scrumPokerProject) : null;
+        return scrumPokerProject != null ? new ProjectSettings(scrumPokerProject) : new ProjectSettings();
     }
 
     /**