+++ docs/changelog.md
@@ -15,6 +15,7 @@ category: Administration
 * chore: optimize imports and code style
 * test: use GitHub actions for continuous integration
 * feat: ensure support for latest Jira releases 9.1.0, 8.20.11
+* fix: project settings cannot be saved without any settings
 
 ### [22.07.0] - 2022-07-11